# Ring: reduce allocations and copies when Merge()-ing states.