--- a/kv/memberlist/mergeable.go
+++ b/kv/memberlist/mergeable.go
@@ -8,6 +8,7 @@ type Mergeable interface {
 	// Merge with other value in place. Returns change, that can be sent to other clients.
 	// If merge doesn't result in any change, returns nil.
 	// Error can be returned if merging with given 'other' value is not possible.
+	// This method may also modify the other parameter.
 	//
 	// In order for state merging to work correctly, Merge function must have some properties. When talking about the
 	// result of the merge in the following text, we don't mean the return value ("change"), but the