**What this PR does**:

Speed up the memberlist receive path by optimizing `ring.Desc.Merge`.

Instead of creating copies of the "Ingesters" maps when normalizing, normalize
them in-place. This reduces the overhead of creating new maps only to discard
the source versions shortly after.

Two additional changes also yielded some improvement:
- When checking for LEFT states, don't update the map unless necessary.
- Collapsing the two loops into one, avoiding iterating the map twice.

Improvement of this change compared to #76 

```
name                             old time/op    new time/op    delta
MemberlistReceiveWithRingDesc-6    1.01ms ± 6%    0.44ms ±15%  -56.56%  (p=0.000 n=9+9)

name                             old alloc/op   new alloc/op   delta
MemberlistReceiveWithRingDesc-6     254kB ± 0%      16kB ± 0%  -93.85%  (p=0.000 n=10+9)

name                             old allocs/op  new allocs/op  delta
MemberlistReceiveWithRingDesc-6       637 ± 0%       612 ± 0%   -3.92%  (p=0.000 n=10+9)
```


Total improvement of **both** #76 and #77

```
name                             old time/op    new time/op    delta
MemberlistReceiveWithRingDesc-6    1.76ms ± 5%    0.44ms ±15%  -74.94%  (p=0.000 n=10+9)

name                             old alloc/op   new alloc/op   delta
MemberlistReceiveWithRingDesc-6     566kB ± 0%      16kB ± 0%  -97.24%  (p=0.000 n=9+9)

name                             old allocs/op  new allocs/op  delta
MemberlistReceiveWithRingDesc-6     1.86k ± 0%     0.61k ± 0%  -67.17%  (p=0.000 n=10+9)
```


<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->



**Which issue(s) this PR fixes**:

Fixes na

**Checklist**
- na Tests updated
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`
    _I will add this PR to the changelog entry for #76 once it's merged, no need for two entries I think._