@bboreham 2021-11-24
Would it be an idea to cache that nothing has changed since the last normalisation?