@stevesg 2021-11-24
> Would it be an idea to cache that nothing has changed since the last normalisation?

That's certainly an option but I'd really like to try and eliminate the normalization entirely, but that's a bit more complex to reason about. I think it still worth committing this in the meanwhile, as it low hanging fruit.